Rename assetutil to aautil
